Move tooltip stories
